Fixed Double Error

+++ BBFitnessTrackerAPI/Core/Contracts/ICalorieDataRepository.cs
@@ -14,19 +14,20 @@ namespace Core.Contracts
 
         Task<CalorieData> GetCalorieDataByIdAsync(int calorieId);
 
-        Task<int> ProteinIntakeByToday(int userId);
-        Task<int> FetteIntakeByToday(int userId);
-        Task<int> CalorieIntakeByDay(int userId, DateTime date);
-        Task<int> KohlenhydrateByDay(int userId, DateTime date);
+        Task<double> ProteinIntakeByToday(int userId);
+        Task<double> FetteIntakeByToday(int userId);
+        Task<double> CalorieIntakeByDay(int userId, DateTime date);
+        Task<double> KohlenhydrateByDay(int userId, DateTime date);
 
-        Task<Dictionary<string, int>> IntakeByWeek(int userId, DateTime startDate);
+        Task<Dictionary<string, double>> IntakeByWeek(int userId, DateTime startDate);
 
-        Task<Dictionary<string, int>> IntakeByMonth(int userId, int year, int month);
+        Task<Dictionary<string, double>> IntakeByMonth(int userId, int year, int month);
 
-        Task<Dictionary<string, int>> IntakeByYear(int userId, int year);
+        Task<Dictionary<string, double>> IntakeByYear(int userId, int year);
 
         Task<Dictionary<int, List<CalorieData>>> GetCalorieDataGroupedByUserId(int userId);
 
+        
 
         void Add(CalorieData calorieData);
         void Update(CalorieData calorieData);
